Small refactoring to polish the PR
