+++ xiplot/app.py
@@ -3,7 +3,10 @@ from collections import Counter
 import dash_mantine_components as dmc
 from dash import ALL, Input, Output, State, ctx, dcc, html
 
-from xiplot.plugin import get_new_plugins_cached, is_plugin_loading_supported
+from xiplot.plugin import (
+    get_new_plugins_cached,
+    is_dynamic_plugin_loading_supported,
+)
 from xiplot.tabs.cluster import Cluster
 from xiplot.tabs.data import Data
 from xiplot.tabs.embedding import Embedding
@@ -31,7 +34,7 @@ class XiPlot:
 
         TABS = [Data, Plots, Cluster, Embedding, Settings]
 
-        if is_plugin_loading_supported():
+        if is_dynamic_plugin_loading_supported():
             TABS.insert(-1, Plugins)
 
         self.app.layout = dmc.NotificationsProvider(
